consolidate everything related to client method return values to ClientMethodsReturnDescription.
