+++ packages/http-client-java/generator/http-client-generator-mgmt/src/main/java/com/microsoft/typespec/http/client/generator/mgmt/mapper/FluentClientMethodMapper.java
@@ -6,6 +6,7 @@ package com.microsoft.typespec.http.client.generator.mgmt.mapper;
 import com.microsoft.typespec.http.client.generator.core.extension.model.codemodel.Operation;
 import com.microsoft.typespec.http.client.generator.core.extension.plugin.JavaSettings;
 import com.microsoft.typespec.http.client.generator.core.mapper.ClientMethodMapper;
+import com.microsoft.typespec.http.client.generator.core.mapper.ClientMethodsReturnDescription;
 import com.microsoft.typespec.http.client.generator.core.model.clientmodel.ClientMethod;
 import com.microsoft.typespec.http.client.generator.core.model.clientmodel.ClientMethodParameter;
 import com.microsoft.typespec.http.client.generator.core.model.clientmodel.ClientMethodType;
@@ -24,7 +25,7 @@ public class FluentClientMethodMapper extends ClientMethodMapper {
 
     @Override
     protected void createAdditionalLroMethods(Operation operation, ClientMethod.Builder builder,
-        List<ClientMethod> methods, boolean isProtocolMethod, ClientMethodsReturnType methodsReturnType,
+        List<ClientMethod> methods, boolean isProtocolMethod, ClientMethodsReturnDescription methodsReturnDescription,
         ProxyMethod proxyMethod, List<ClientMethodParameter> parameters,
         boolean generateClientMethodWithOnlyRequiredParameters, MethodOverloadType defaultOverloadType) {
 
@@ -33,7 +34,7 @@ public class FluentClientMethodMapper extends ClientMethodMapper {
         // API
 
         // async
-        methods.add(builder.returnValue(methodsReturnType.getFor(ClientMethodType.LongRunningAsync))
+        methods.add(builder.returnValue(methodsReturnDescription.getReturnValue(ClientMethodType.LongRunningAsync))
             .name(proxyMethod.getSimpleAsyncMethodName())
             .onlyRequiredParameters(false)
             .type(ClientMethodType.LongRunningAsync)
@@ -55,7 +56,7 @@ public class FluentClientMethodMapper extends ClientMethodMapper {
             parameters, getContextParameter(isProtocolMethod));
 
         // sync
-        methods.add(builder.returnValue(methodsReturnType.getFor(ClientMethodType.LongRunningSync))
+        methods.add(builder.returnValue(methodsReturnDescription.getReturnValue(ClientMethodType.LongRunningSync))
             .name(proxyMethod.getName())
             .onlyRequiredParameters(false)
             .type(ClientMethodType.LongRunningSync)
